requirements.txt
arxiv
duckduckgo-search
faiss-cpu
langchain
langchain-openai
langgraph
pymupdf
python-dotenv